refactor: refactor entities
