+++ src/web_socket_server/index.ts
@@ -1,18 +1,5 @@
 import WebSocket, { WebSocketServer } from "ws";
-import {
-  mouse,
-  left,
-  right,
-  up,
-  down,
-  straightTo,
-  screen,
-  Image,
-  imageResource,
-} from "@nut-tree/nut-js";
-import COMMANDS from "../constants/commands";
-import { drawCircle, drawRectangle } from "../drawing/drawing";
-import { moveDown, moveLeft, moveRight, moveUp } from "../navigation/navigation";
+import { handleWsData } from "./entities";
 
 // Creating connection using websocket
 
@@ -24,37 +11,11 @@ const onConnectWS = async (ws: WebSocket) => {
     ws.send(data.toString());
     console.log(JSON.stringify(data));
     console.log(`Client has sent us: ${data.toString("utf8")}`);
+    handleWsData(data)
 
-    const [command, _width, _length] = data.toString().split(" ");
+ 
 
-    const width = Number(_width);
-    const length = Number(_length);
-    if (command === COMMANDS.PRNT) {
-      let i;
-      //    const img = await screen.capture("./").then((im) => ws.send(im));
-    }
-
-    if (command === COMMANDS.UP) {
-      await moveUp(width)
-    }
-    if (command === COMMANDS.DOWN) {
-      await moveDown(width)
-    }
-    if (command === COMMANDS.LEFT) {
-      await moveLeft(width)
-    }
-    if (command === COMMANDS.RIGHT) {
-      await moveRight(width)
-    }
-    if (command === COMMANDS.SQUARE) {
-      await drawRectangle(width, width);
-    }
-    if (command === COMMANDS.REC) {
-      await drawRectangle(width, length);
-    }
-    if (command === COMMANDS.CIRCLE) {
-      await drawCircle(width);
-    }
+    
   });
   // handling what to do when clients disconnects from server
   ws.on("close", () => {
